patch-automation/requirements.txt
# Core dependencies
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0

# XML processing
lxml>=4.9.0

# Extended file attributes (for caching)
xattr>=0.10.0

# System monitoring (optional)
psutil>=5.9.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# For better JSON handling
jsonschema>=4.19.0